docnest_backend/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic[email]==2.5.3
python-dotenv==1.0.0
boto3==1.34.14  # For S3 file storage
python-magic==0.4.27  # For file type validation